requirements.txt
cffi==1.15.0
cryptography==36.0.0
Django==3.2.9
dill==0.3.4
django-cors-headers==3.10.0
django-filter==21.1
django-rest-knox==4.1.0
djangorestframework==3.12.4
numpy==1.21.4
pandas==1.3.5
Pillow==9.0.0
pycparser==2.21
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
threadpoolctl==3.0.0
asgiref==3.5.0
django-heroku==0.3.1
gunicorn==20.1.0
psycopg2==2.9.3
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.0.0
